Adding refactoring JqueryuiDroppablePage.java, project structure and testng
