+++ src/test/java/pages/google/GoogleHomePage.java
@@ -1,8 +1,9 @@
-package pages;
+package pages.google;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
+import pages.AbstractPage;
 
 public class GoogleHomePage extends AbstractPage {
