+++ src/test/java/pages/google/GoogleSearchResultsPage.java
@@ -1,9 +1,10 @@
-package pages;
+package pages.google;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindAll;
 import org.openqa.selenium.support.FindBy;
+import pages.AbstractPage;
 
 import java.util.List;
